Переместил функцию осздания окна с ошибкой
