+++ js/main.js
@@ -1,6 +1,6 @@
 import {getOffersData} from './api.js';
 import {placeAllOffersOnMap} from './map.js';
-import {makeFailureMessage} from './utils.js';
+import {makeFailureMessage} from './form';
 import './card.js';
 import './form.js';
